Several small bugfixes and updates.
Added minElements and maxElements to freemarker objects (they were
    present in the data, but not being pushed into freemarker).

Fixed display names not to show up with '-' as values when they are
empty or null.

ProcessIndexTemplate is now protected.

In the default templates, arguments that have no synonyms no longer
render synonyms at all (as opposed to showing up as NA).